+++ examples/with-secret/.gitignore
@@ -1,2 +1,3 @@
 .kubricate
 output
+node_modules
