+++ examples/with-secret/src/stack-templates/cronJobTemplate.ts
